Issue 79: Move controllers to root of folder
